requirements.txt
requests
python-dotenv
textual>=0.50.0
